requirements.txt
fastapi
uvicorn
dotenv
openai
pydantic
pdfminer.six
pymupdf
python-multipart
asyncio
pandas
openpyxl
xlrd
